read/write return ssize_t not int

+++ user/_read.c
@@ -1,3 +1,3 @@
 #include "scaraOS.h"
 
-_SYSCALL3(_SYS_read, int, _read, int, void *, size_t);
+_SYSCALL3(_SYS_read, ssize_t, _read, int, void *, size_t);
