+++ user/_write.c
@@ -1,3 +1,3 @@
 #include "scaraOS.h"
 
-_SYSCALL3(_SYS_write, int, _write, int, const void *, size_t);
+_SYSCALL3(_SYS_write, ssize_t, _write, int, const void *, size_t);
